requirements.txt
github-copilot-sdk>=0.1.0
playwright>=1.49.0
pydantic>=2.0.0
rich>=13.0.0
